+++ packages/backend/index.ts
@@ -1,14 +1,19 @@
 import fastify from "fastify";
 import fastifyStatic from "fastify-static";
-import mercurius from "mercurius";
+import mercurius, { MercuriusContext } from "mercurius";
 import { schema } from "@packages/schema";
 import { config } from "@packages/config";
+
+export type Context = MercuriusContext;
 
 export const app = fastify();
 
 // Mercurius GraphQL server
 app.register(mercurius, {
   schema,
+  context: () => {
+    return {};
+  },
   subscription: true,
   graphiql: config.mode === "development" && "playground",
 });